Create arrange.js
I created this script to automate the rearranging process I have to do each week to a Usage Report I get from our server team on all of our apps. It takes quite a bit of manual manipulating and sorting to get it in the order I like, so I finally wrote a script to do it.